+++ app/auth/templates/auth/register.html
@@ -1,27 +1,44 @@
 {% extends "main/base.html" %}
-{% block content %}
-    <h2>Register</h2>
-    <form method="POST">
-        {{ form.hidden_tag() }}
-        <div>
-            {{ form.username.label }}
-            {{ form.username(size=32) }}
-        </div>
-        <div>
-            {{ form.email.label }}
-            {{ form.email(size=32) }}
-        </div>
-        <div>
-            {{ form.password.label }}
-            {{ form.password(size=32) }}
-        </div>
-        <div>
-            {{ form.confirm_password.label }}
-            {{ form.confirm_password(size=32) }}
-        </div>
-        <div>
-            {{ form.submit() }}
-        </div>
-    </form>
-    <a href="{{ url_for('login') }}">Already have an account? Login here.</a>
+
+{% block title %}Register{% endblock %}
+
+{% block pagebody%}
+<div id="smallwindow">
+    <div class="titlebar">
+        <h2>Register</h2>
+    </div>
+    <div class="smallcontent">
+        <form method="POST" action="{{ url_for('auth.register')}}">
+            {{ form.hidden_tag() }}
+            <div>
+                {{ form.first_name.label }}
+                {{ form.first_name(size=32) }}
+            </div>
+            <div>
+                {{ form.last_name.label }}
+                {{ form.last_name(size=32) }}
+            </div>
+            <div>
+                {{ form.username.label }}
+                {{ form.username(size=32) }}
+            </div>
+            <div>
+                {{ form.email.label }}
+                {{ form.email(size=32) }}
+            </div>
+            <div>
+                {{ form.password.label }}
+                {{ form.password(size=32) }}
+            </div>
+            <div>
+                {{ form.confirm_password.label }}
+                {{ form.confirm_password(size=32) }}
+            </div>
+            <div>
+                {{ form.submit() }}
+            </div>
+        </form>
+        <a href="{{ url_for('auth.login') }}">Already have an account? Login here.</a>
+    </div>
+</div>
 {% endblock %}
