--- app/instance/models/__init__.py
@@ -1,1 +1,0 @@
-
